Docker/gammrf/requirements.txt
iniparse==0.4
gps3==0.33.3
numpy==1.12.0
matplotlib==2.0.0
pyrtlsdr==0.2.4
pyModeS==1.0.7
pyusb==1.0.0
